requirements copy.txt
cat > requirements.txt << 'EOF'
python-telegram-bot==21.5
python-dotenv
requests
yt-dlp
beautifulsoup4